safety_controller:
This node takes in a desired velocity as a geometry_msgs/Twist-message
and laser data as sensor_msgs/LaserScan.
With the laser data a costmap_2d is build up.
The costmap in combination to a base_local_planner instance is able to
check whether a desired velocity is feasible (collision free) or not.
If not, the velocity commands in linear.x and linear.y are multiplied by
0.5 in order to limit the velocity.