add feature to increase or decrease velocities for planner
This feature allows to setup velocities based on laser range data.
It also incorporates the current costmap representation in terms of the
cost represented at the current robot's pose.
In case the cost at the current pose is higher than a certain threshold
it tells the local planner of the move_base to set a lower velocity.
A laser is taken into account by setting a lower velocity in case one of
the laser beams represents an obstacle closer than one meter.